// src/app/modules/shopSettings/shopSettings.interface.ts
export interface IShopSettings {
  websiteName: string;
  websiteTitle: string;
  logo: string;
  email: string;
  phone: string;
  address: string;
  mapEmbedUrl?: string;
  aboutUs: string;      
  ourStory: string;     
  privacyPolicy: string;
  termsConditions: string;
  refundPolicy: string;
  facebookUrl?: string;
  instagramUrl?: string;
  youtubeUrl?: string;
  linkedinUrl?: string;
}